pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doc-finder"
version = "0.1.0"
description = "Minimal LangGraph scaffold for doc-finder"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "accelerate>=1.13,<2.0",
    "cairosvg>=2.7,<3.0",
    "einops>=0.8,<1.0",
    "fastapi>=0.115,<1.0",
    "langgraph>=1.0,<2.0",
    "peft>=0.19,<1.0",
    "pillow>=10.0,<12.0",
    "psycopg[binary]>=3.2,<4.0",
    "pyyaml>=6.0,<7.0",
    "timm>=1.0,<2.0",
    "torch>=2.3,<3.0",
    "transformers>=4.46,<5.0",
    "uvicorn>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
